StringBuilder capacity bug
Originally the StringBuilder's capacity was half what it was suppose to be. SequenceLength represents the number of ATCGs in a sequence. The StringBuilder stores two numbers for every letter in a sequence (example: '00' are stored in place of an 'A'). This means that the StringBuilder's capacity needs to be twice the sequenceLength so that it can store all the zeros and ones for all the ATCGs in a sequence. Example, if a sequence was 'ATACG', StringBuilder would store '0011000110', which is twice the length of the original sequence.

I also added some TODOs that list possible bugs or improvements in the code.